TLRC - 3824 - Printing Integration - Integrate with PrintRequest - TLRC - 3851 - Dev - Added Integration tests for GetPendingPrintRequestsAsync and UpdatePrintReqeustResponsesAsync methods

+++ src/Sfa.Tl.ResultsAndCertification.Models/Contracts/Printing/Certificate.cs
@@ -4,7 +4,7 @@
     {
         public string CertificateNumber { get; set; }
         public string Type { get; set; }
-        public int Uln { get; set; }
+        public long Uln { get; set; }
         public string LearnerName { get; set; }
         public LearningDetails LearningDetails { get; set; }
     }
